Changed the return types for the size function to int

+++ include/FTSize.h
@@ -14,11 +14,11 @@ class	FTSize
 		FTSize();
 		virtual ~FTSize();
 		bool CharSize( FT_Face* face, unsigned int point_size, unsigned int x_resolution, unsigned int y_resolution );
-		float Ascender() const;
-		float Descender() const;
-		float Height() const;
-		float Width() const;
-		float Underline() const;
+		int Ascender() const;
+		int Descender() const;
+		int Height() const;
+		int Width() const;
+		int Underline() const;
 
 		FT_Error Error() const { return err; }
